+++ src/main/java/com/ecommerce/backend/repository/ProductRepository.java
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
     boolean existsProductByName(String name);
 
-    boolean existsProductById(Long id);
+    boolean existsProductById(Integer id);
 
 }
